feat: add explicit build
