+++ src/commands/deploy.ts
@@ -18,7 +18,7 @@ import fsPromises from "node:fs/promises";
 import path from "node:path";
 
 import color from "@oclif/color";
-import { CliUx, Command, Flags } from "@oclif/core";
+import { Command, Flags } from "@oclif/core";
 import { yamlDiffPatch } from "yaml-diff-patch";
 
 import { AquaCLI, initAquaCli } from "../lib/aquaCli";
@@ -28,25 +28,8 @@ import {
   initNewReadonlyAppConfig,
   ServicesV3,
 } from "../lib/configs/project/app";
-import {
-  Distribution,
-  DISTRIBUTION_EVEN,
-  FluenceConfigReadonly,
-  OverrideModules,
-  ServiceDeployV1,
-} from "../lib/configs/project/fluence";
-import {
-  initReadonlyModuleConfig,
-  ModuleConfigReadonly,
-  MODULE_TYPE_RUST,
-} from "../lib/configs/project/module";
-import { initProjectSecretsConfig } from "../lib/configs/project/projectSecrets";
-import {
-  FACADE_MODULE_NAME,
-  initReadonlyServiceConfig,
-  ModuleV0,
-  ServiceConfigReadonly,
-} from "../lib/configs/project/service";
+import type { ModuleConfigReadonly } from "../lib/configs/project/module";
+import type { ModuleV0 } from "../lib/configs/project/service";
 import {
   AQUA_EXT,
   CommandObj,
@@ -56,9 +39,7 @@ import {
   FS_OPTIONS,
   KEY_PAIR_FLAG,
   KEY_PAIR_FLAG_NAME,
-  MODULE_CONFIG_FILE_NAME,
   NO_INPUT_FLAG,
-  SERVICE_CONFIG_FILE_NAME,
   TIMEOUT_FLAG,
   TIMEOUT_FLAG_NAME,
 } from "../lib/const";
@@ -67,34 +48,17 @@ import {
   generateRegisterApp,
 } from "../lib/deployedApp";
 import {
-  downloadModule,
-  downloadService,
   getModuleUrlOrAbsolutePath,
   getModuleWasmPath,
-  isUrl,
-  validateAquaName,
 } from "../lib/helpers/downloadFile";
 import { ensureFluenceProject } from "../lib/helpers/ensureFluenceProject";
 import { generateServiceInterface } from "../lib/helpers/generateServiceInterface";
 import { getIsInteractive } from "../lib/helpers/getIsInteractive";
 import { getMessageWithKeyValuePairs } from "../lib/helpers/getMessageWithKeyValuePairs";
 import { replaceHomeDir } from "../lib/helpers/replaceHomeDir";
-import {
-  ConfigKeyPair,
-  generateKeyPair,
-  getExistingKeyPair,
-  getProjectKeyPair,
-  getUserKeyPair,
-} from "../lib/keypairs";
-import { initMarineCli, MarineCLI } from "../lib/marineCli";
-import {
-  getEvenlyDistributedIds,
-  getEvenlyDistributedIdsFromTheList,
-  getRandomRelayAddr,
-  getRandomRelayId,
-  getRandomRelayIdFromTheList,
-  Relays,
-} from "../lib/multiaddr";
+import { ConfigKeyPair, getExistingKeyPair } from "../lib/keypairs";
+import { initMarineCli } from "../lib/marineCli";
+import { getRandomRelayAddr } from "../lib/multiaddr";
 import {
   ensureFluenceAquaServicesDir,
   ensureFluenceTmpDeployJsonPath,
@@ -102,6 +66,7 @@ import {
 import { confirm } from "../lib/prompt";
 import { hasKey } from "../lib/typeHelpers";
 
+import { build, resolveServiceInfos } from "./build";
 import { removeApp } from "./remove";
 
 export default class Deploy extends Command {
@@ -313,268 +278,30 @@ type DeployInfo = {
   keyPair: ConfigKeyPair;
 };
 
-const overrideModule = (
-  mod: ModuleV0,
-  overrideModules: OverrideModules | undefined,
-  moduleName: string
-): ModuleV0 => ({ ...mod, ...overrideModules?.[moduleName] });
-
 type PreparedForDeploy = Omit<DeployInfo, "modules" | "serviceDirPath"> & {
   deployJSON: DeployJSONConfig;
   facadeModuleWasmPath: string;
 };
 
-type PrepareForDeployArg = {
-  commandObj: CommandObj;
-  fluenceConfig: FluenceConfigReadonly;
-  defaultKeyPair: ConfigKeyPair;
-  isInteractive: boolean;
-  marineCli: MarineCLI;
-};
+type PrepareForDeployArg = Omit<
+  Parameters<typeof resolveServiceInfos>[0] & Parameters<typeof build>[0],
+  "allServiceInfos"
+>;
 
 const prepareForDeploy = async ({
-  commandObj,
-  fluenceConfig,
-  defaultKeyPair,
-  isInteractive,
   marineCli,
+  commandObj,
+  ...rest
 }: PrepareForDeployArg): Promise<Array<PreparedForDeploy>> => {
-  if (
-    fluenceConfig.services === undefined ||
-    Object.keys(fluenceConfig.services).length === 0
-  ) {
-    throw new Error(
-      `Use ${color.yellow(
-        "fluence service add"
-      )} command to add services you want to deploy to ${color.yellow(
-        FLUENCE_CONFIG_FILE_NAME
-      )}`
-    );
-  }
+  const allServiceInfos = await resolveServiceInfos({ commandObj, ...rest });
 
-  type ServicePathPromises = Promise<{
-    serviceName: string;
-    serviceDirPath: string;
-    get: string;
-    deploy: Array<ServiceDeployV1>;
-    keyPair: ConfigKeyPair;
-  }>;
+  const mapOfAllModuleConfigs = await build({
+    allServiceInfos,
+    commandObj,
+    marineCli,
+  });
 
-  CliUx.ux.action.start("Making sure all services are downloaded");
-
-  const projectSecretsConfig = await initProjectSecretsConfig(commandObj);
-
-  const getKeyPair = async (
-    defaultKeyPair: ConfigKeyPair,
-    keyPairName: string | undefined
-  ): Promise<ConfigKeyPair> => {
-    if (keyPairName === undefined) {
-      return defaultKeyPair;
-    }
-
-    let keyPair =
-      (await getProjectKeyPair({
-        commandObj,
-        keyPairName,
-      })) ??
-      (await getUserKeyPair({
-        commandObj,
-        keyPairName,
-      }));
-
-    if (keyPair === undefined) {
-      CliUx.ux.action.stop("paused");
-
-      commandObj.warn(`Key pair ${color.yellow(keyPairName)} not found`);
-
-      const doGenerate = await confirm({
-        message: `Do you want to generate new key-pair ${color.yellow(
-          keyPairName
-        )} for your project?`,
-        isInteractive,
-      });
-
-      if (!doGenerate) {
-        return commandObj.error("Aborted");
-      }
-
-      CliUx.ux.action.start("Making sure all services are downloaded");
-
-      keyPair = await generateKeyPair(keyPairName);
-      projectSecretsConfig.keyPairs.push(keyPair);
-      await projectSecretsConfig.$commit();
-    }
-
-    return keyPair;
-  };
-
-  const servicePaths = await Promise.all(
-    Object.entries(fluenceConfig.services).map(
-      ([serviceName, { get, deploy, keyPairName }]): ServicePathPromises =>
-        (async (): ServicePathPromises => ({
-          serviceName,
-          deploy,
-          get,
-          keyPair: await getKeyPair(defaultKeyPair, keyPairName),
-          serviceDirPath: isUrl(get)
-            ? await downloadService(get)
-            : path.resolve(get),
-        }))()
-    )
-  );
-
-  CliUx.ux.action.stop();
-
-  type ServiceConfigPromises = Promise<{
-    serviceName: string;
-    serviceConfig: ServiceConfigReadonly;
-    serviceDirPath: string;
-    deploy: Array<ServiceDeployV1>;
-    keyPair: ConfigKeyPair;
-  }>;
-
-  const serviceConfigs = await Promise.all(
-    servicePaths.map(
-      ({
-        serviceName,
-        serviceDirPath,
-        deploy,
-        get,
-        keyPair,
-      }): ServiceConfigPromises =>
-        (async (): ServiceConfigPromises => ({
-          serviceName,
-          deploy,
-          keyPair,
-          serviceConfig:
-            (await initReadonlyServiceConfig(serviceDirPath, commandObj)) ??
-            commandObj.error(
-              `Service ${color.yellow(serviceName)} must have ${color.yellow(
-                SERVICE_CONFIG_FILE_NAME
-              )}. ${
-                isUrl(get)
-                  ? `Not able to find it after downloading and decompressing ${color.yellow(
-                      get
-                    )}`
-                  : `Not able to find it at ${color.yellow(get)}`
-              }`
-            ),
-          serviceDirPath,
-        }))()
-    )
-  );
-
-  const allDeployInfos = await Promise.all(
-    serviceConfigs.flatMap(
-      ({
-        serviceName,
-        deploy,
-        serviceConfig,
-        serviceDirPath,
-        keyPair,
-      }): Array<Promise<DeployInfo>> =>
-        deploy.flatMap(
-          ({
-            deployId,
-            count,
-            peerId,
-            peerIds,
-            overrideModules,
-            distribution,
-            keyPairName,
-          }): Array<Promise<DeployInfo>> => {
-            const deployIdValidity = validateAquaName(deployId);
-
-            if (deployIdValidity !== true) {
-              return commandObj.error(
-                `deployId ${color.yellow(deployId)} ${deployIdValidity}`
-              );
-            }
-
-            return getPeerIds({
-              peerId: peerIds ?? peerId,
-              distribution,
-              count,
-              relays: fluenceConfig.relays,
-              namedPeerIds: fluenceConfig.peerIds,
-            }).map(
-              (peerId: string): Promise<DeployInfo> =>
-                (async (): Promise<DeployInfo> => ({
-                  serviceName,
-                  serviceDirPath,
-                  deployId,
-                  peerId:
-                    typeof peerId === "string"
-                      ? fluenceConfig?.peerIds?.[peerId ?? ""] ?? peerId
-                      : peerId,
-                  modules: getDeployInfoModules({
-                    commandObj,
-                    deployId,
-                    overrideModules,
-                    serviceConfigModules: serviceConfig.modules,
-                    serviceDirPath,
-                    serviceName,
-                  }),
-                  keyPair: await getKeyPair(keyPair, keyPairName),
-                }))()
-            );
-          }
-        )
-    )
-  );
-
-  const setOfAllGets = [
-    ...new Set(
-      allDeployInfos.flatMap(
-        ({ modules, serviceDirPath }): Array<string> =>
-          modules.map(({ moduleConfig: { get } }): string =>
-            getModuleUrlOrAbsolutePath(get, serviceDirPath)
-          )
-      )
-    ),
-  ];
-
-  CliUx.ux.action.start("Making sure all modules are downloaded and built");
-
-  const mapOfAllModuleConfigs = new Map(
-    await Promise.all(
-      setOfAllGets.map(
-        (get): Promise<[string, ModuleConfigReadonly]> =>
-          (async (): Promise<[string, ModuleConfigReadonly]> => {
-            const moduleConfig = isUrl(get)
-              ? await initReadonlyModuleConfig(
-                  await downloadModule(get),
-                  commandObj
-                )
-              : await initReadonlyModuleConfig(get, commandObj);
-
-            if (moduleConfig === null) {
-              CliUx.ux.action.stop(color.red("error"));
-
-              return commandObj.error(
-                `Module with get: ${color.yellow(
-                  get
-                )} doesn't have ${color.yellow(MODULE_CONFIG_FILE_NAME)}`
-              );
-            }
-
-            if (moduleConfig.type === MODULE_TYPE_RUST) {
-              await marineCli({
-                args: ["build"],
-                flags: { release: true },
-                cwd: path.dirname(moduleConfig.$getPath()),
-              });
-            }
-
-            return [get, moduleConfig];
-          })()
-      )
-    )
-  );
-
-  CliUx.ux.action.stop();
-
-  return allDeployInfos.map(
+  return allServiceInfos.map(
     ({ modules, serviceDirPath, ...rest }): PreparedForDeploy => {
       const moduleConfigs = modules.map(
         ({
@@ -623,109 +350,6 @@ const prepareForDeploy = async ({
       };
     }
   );
-};
-
-type GetPeerIdsArg = {
-  peerId: undefined | string | Array<string>;
-  distribution: Distribution | undefined;
-  count: number | undefined;
-  relays: Relays;
-  namedPeerIds: Record<string, string> | undefined;
-};
-
-const getNamedPeerIdsFn =
-  (
-    namedPeerIds: Record<string, string>
-  ): ((peerIds: Array<string>) => string[]) =>
-  (peerIds: Array<string>): string[] =>
-    peerIds.map((peerId): string => namedPeerIds[peerId] ?? peerId);
-
-const getPeerIds = ({
-  peerId,
-  distribution = DISTRIBUTION_EVEN,
-  count,
-  relays,
-  namedPeerIds = {},
-}: GetPeerIdsArg): Array<string> => {
-  const getNamedPeerIds = getNamedPeerIdsFn(namedPeerIds);
-
-  if (distribution === DISTRIBUTION_EVEN) {
-    if (peerId === undefined) {
-      return getEvenlyDistributedIds(relays, count);
-    }
-
-    return getEvenlyDistributedIdsFromTheList(
-      getNamedPeerIds(typeof peerId === "string" ? [peerId] : peerId),
-      count
-    );
-  }
-
-  if (peerId === undefined) {
-    return Array.from({ length: count ?? 1 }).map((): string =>
-      getRandomRelayId(relays)
-    );
-  }
-
-  const peerIds = typeof peerId === "string" ? [peerId] : peerId;
-  return Array.from({ length: count ?? peerIds.length }).map((): string =>
-    getRandomRelayIdFromTheList(peerIds)
-  );
-};
-
-type GetDeployInfoModulesArg = {
-  commandObj: CommandObj;
-  overrideModules: OverrideModules | undefined;
-  serviceName: string;
-  deployId: string;
-  serviceDirPath: string;
-  serviceConfigModules: { facade: ModuleV0 } & Record<string, ModuleV0>;
-};
-
-const getDeployInfoModules = ({
-  commandObj,
-  overrideModules,
-  serviceName,
-  deployId,
-  serviceDirPath,
-  serviceConfigModules,
-}: GetDeployInfoModulesArg): Array<DeployInfoModule> => {
-  const modulesNotFoundInServiceYaml = Object.keys(
-    overrideModules ?? {}
-  ).filter((moduleName): boolean => !(moduleName in serviceConfigModules));
-
-  if (modulesNotFoundInServiceYaml.length > 0) {
-    commandObj.error(
-      `${color.yellow(FLUENCE_CONFIG_FILE_NAME)} has service ${color.yellow(
-        serviceName
-      )} with deployId ${color.yellow(
-        deployId
-      )} that has moduleOverrides for modules that don't exist in the service ${color.yellow(
-        serviceDirPath
-      )}. Please make sure ${color.yellow(
-        modulesNotFoundInServiceYaml.join(", ")
-      )} spelled correctly `
-    );
-  }
-
-  const { [FACADE_MODULE_NAME]: facadeModule, ...otherModules } =
-    serviceConfigModules;
-
-  return [
-    ...Object.entries(otherModules).map(
-      ([moduleName, mod]): DeployInfoModule => ({
-        moduleConfig: overrideModule(mod, overrideModules, moduleName),
-        moduleName,
-      })
-    ),
-    {
-      moduleConfig: overrideModule(
-        facadeModule,
-        overrideModules,
-        FACADE_MODULE_NAME
-      ),
-      moduleName: FACADE_MODULE_NAME,
-    },
-  ];
 };
 
 /* eslint-disable camelcase */
